move retrieve public workflow to hub resource
