+++ core/gui/src/app/hub/service/workflow/hub-workflow.service.ts
@@ -4,6 +4,9 @@ import { Observable } from "rxjs";
 import { AppSettings } from "../../../common/app-setting";
 import { HubWorkflow } from "../../component/type/hub-workflow.interface";
 import { User } from "src/app/common/type/user";
+import { Workflow } from "../../../common/type/workflow";
+import { filter, map } from "rxjs/operators";
+import { WorkflowUtilService } from "../../../workspace/service/workflow-graph/util/workflow-util.service";
 
 @Injectable({
   providedIn: "root",
@@ -29,5 +32,12 @@ export class HubWorkflowService {
   public getWorkflowName(wid: number): Observable<string> {
     const params = new HttpParams().set("wid", wid);
     return this.http.get(`${this.BASE_URL}/workflow_name/`, { params, responseType: "text" });
+  }
+
+  public retrievePublicWorkflow(wid: number): Observable<Workflow> {
+    return this.http.get<Workflow>(`${this.BASE_URL}/public/${wid}`).pipe(
+      filter((workflow: Workflow) => workflow != null),
+      map(WorkflowUtilService.parseWorkflowInfo)
+    );
   }
 }
